refactor: split session transcript handlers
