+++ internal/api/handler_session_chat.go
@@ -8,7 +8,6 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"strconv"
 	"strings"
 
 	"github.com/gastownhall/gascity/internal/beads"
@@ -33,22 +32,6 @@ type sessionCreateRequest struct {
 	// Stored in bead metadata for session-to-project association.
 	ProjectID string `json:"project_id,omitempty"`
 	Title     string `json:"title,omitempty"`
-}
-
-type sessionTranscriptResponse struct {
-	ID         string                       `json:"id"`
-	Template   string                       `json:"template"`
-	Format     string                       `json:"format"`
-	Turns      []outputTurn                 `json:"turns"`
-	Pagination *worker.TranscriptPagination `json:"pagination,omitempty"`
-}
-
-type sessionRawTranscriptResponse struct {
-	ID         string                       `json:"id"`
-	Template   string                       `json:"template"`
-	Format     string                       `json:"format"`
-	Messages   []json.RawMessage            `json:"messages"`
-	Pagination *worker.TranscriptPagination `json:"pagination,omitempty"`
 }
 
 func writeSessionManagerError(w http.ResponseWriter, err error) {
@@ -461,134 +444,4 @@ func (s *Server) persistSessionMeta(store beads.Store, sessionID, kind, projectI
 			log.Printf("persistSessionMeta: session %s: %v", sessionID, err)
 		}
 	}
-}
-
-func (s *Server) handleSessionTranscript(w http.ResponseWriter, r *http.Request) {
-	store := s.state.CityBeadStore()
-	if store == nil {
-		writeError(w, http.StatusServiceUnavailable, "unavailable", "no bead store configured")
-		return
-	}
-
-	id, err := s.resolveSessionIDAllowClosedWithConfig(store, r.PathValue("id"))
-	if err != nil {
-		writeResolveError(w, err)
-		return
-	}
-
-	catalog, err := s.workerSessionCatalog(store)
-	if err != nil {
-		writeSessionManagerError(w, err)
-		return
-	}
-	info, err := catalog.Get(id)
-	if err != nil {
-		writeSessionManagerError(w, err)
-		return
-	}
-	handle, err := s.workerHandleForSession(store, id)
-	if err != nil {
-		writeSessionManagerError(w, err)
-		return
-	}
-	path, err := handle.TranscriptPath(r.Context())
-	if err != nil && !errors.Is(err, worker.ErrHistoryUnavailable) {
-		writeSessionManagerError(w, err)
-		return
-	}
-
-	wantRaw := r.URL.Query().Get("format") == "raw"
-
-	if path != "" {
-		tail := 0
-		if v := r.URL.Query().Get("tail"); v != "" {
-			if n, convErr := strconv.Atoi(v); convErr == nil && n >= 0 {
-				tail = n
-			}
-		}
-		before := r.URL.Query().Get("before")
-
-		if wantRaw {
-			transcript, err := handle.Transcript(r.Context(), worker.TranscriptRequest{
-				TailCompactions: tail,
-				BeforeEntryID:   before,
-				Raw:             true,
-			})
-			if err != nil {
-				writeError(w, http.StatusInternalServerError, "internal", "reading session log: "+err.Error())
-				return
-			}
-			writeJSON(w, http.StatusOK, sessionRawTranscriptResponse{
-				ID:         info.ID,
-				Template:   info.Template,
-				Format:     "raw",
-				Messages:   transcript.RawMessages,
-				Pagination: transcript.Session.Pagination,
-			})
-			return
-		}
-
-		transcript, err := handle.Transcript(r.Context(), worker.TranscriptRequest{
-			TailCompactions: tail,
-			BeforeEntryID:   before,
-		})
-		if err != nil {
-			writeError(w, http.StatusInternalServerError, "internal", "reading session log: "+err.Error())
-			return
-		}
-		sess := transcript.Session
-
-		turns := make([]outputTurn, 0, len(sess.Messages))
-		for _, entry := range sess.Messages {
-			turn := entryToTurn(entry)
-			if turn.Text == "" {
-				continue
-			}
-			turns = append(turns, turn)
-		}
-		writeJSON(w, http.StatusOK, sessionTranscriptResponse{
-			ID:         info.ID,
-			Template:   info.Template,
-			Format:     "conversation",
-			Turns:      turns,
-			Pagination: sess.Pagination,
-		})
-		return
-	}
-
-	if wantRaw {
-		writeJSON(w, http.StatusOK, sessionRawTranscriptResponse{
-			ID:       info.ID,
-			Template: info.Template,
-			Format:   "raw",
-			Messages: []json.RawMessage{},
-		})
-		return
-	}
-
-	output, peekErr := handle.Peek(r.Context(), 100)
-	if peekErr != nil && !errors.Is(peekErr, session.ErrSessionInactive) {
-		writeError(w, http.StatusInternalServerError, "internal", peekErr.Error())
-		return
-	}
-	if peekErr == nil {
-		turns := []outputTurn{}
-		if output != "" {
-			turns = append(turns, outputTurn{Role: "output", Text: output})
-		}
-		writeJSON(w, http.StatusOK, sessionTranscriptResponse{
-			ID:       info.ID,
-			Template: info.Template,
-			Format:   "text",
-			Turns:    turns,
-		})
-		return
-	}
-
-	writeJSON(w, http.StatusOK, sessionTranscriptResponse{
-		ID:       info.ID,
-		Template: info.Template,
-		Format:   "conversation",
-		Turns:    []outputTurn{},
-	})
 }
